Make Registration implement an interface as well
